requirements.txt
google-generativeai
Flask
requests
pytest
